refactor(audience): extract shared integration-logs constants
